fix: Detect FreeRDP even if a binary with higher priority but a non-3.x.x version is installed
Previously, there was an edge case where for example the Flatpak variant was installed, but the `xfreerdp` binary with higher priority was also present, however the version was <3.x.x and the specs check rejected it. This commit should hopefully fix such edge cases since we simply return the binary matching 3.x.x and specs.ts is no longer responsible for version checking.

Co-authored-by: Levev <levvevv@gmail.com>